Download Button UI component in download queue * Bumped script memory limit to 4g in php.ini to help with unpacking large files. * Added meta, file_name and media_type to Download object in instances were a download object was created. * Added meta, file_name and media_type to API resources concerned with displaying download objects. * Fixed a fatal bug in RemoveCompletedDownload job where the user might delete a file before the job attempts to. * Added exception logging in File Transfer job. * In the download queue drawer UI, put the file save download button into its own component.   * Can be used in any context where a download object and settings are supplied as props. * Added saveDownloadDestination method to Browse page, makes call to the API to designate a download destination for a download object. * Fixed a bug in the Browse page UI where the Download Queue Drawer was covering the bottom part of the page content. * Added Axios configuration to use maxRedirects = 1, so axios at least follows 1 redirect but no more. * Fixed a bug in the mediaTypeToStoreMap mapping. * Changed UI functions to use download.media_type instead of download.packet.media_type   * The packet can change its contents, but the download columns are frozen on the file that was downloaded. * Changed UI functions to use download.meta instead of download.packet.meta

+++ src/app/Jobs/RemoveCompletedDownload.php
@@ -40,7 +40,9 @@ class RemoveCompletedDownload implements ShouldQueue, ShouldBeUnique
     public function handle(): void
     {
         $this->releaseLock();
-        unlink($this->download->file_uri);
+        if (file_exists($this->download->file_uri)) {
+            unlink($this->download->file_uri);
+        }
 
         // Move the download to the archives table
         ArchiveDownload::dispatch($this->download);
